+++ cg/meta/upload/gisaid/gisaid.py
@@ -1,8 +1,12 @@
-"""Interactions with the gisaid cli upload"""
-
+"""Interactions with the gisaid cli upload_results_to_gisaid"""
+import json
 import logging
 from pathlib import Path
-from typing import List, Optional
+from typing import List, Dict, Optional
+import pandas as pd
+
+from housekeeper.store.models import File
+import tempfile
 
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
@@ -10,11 +14,11 @@ from cg.models.cg_config import CGConfig
 from cg.store import models, Store
 from cg.utils import Process
 from .constants import HEADERS
-from .models import GisaidSample, FastaFile
-from housekeeper.store import models as hk_models
-import csv
+from .models import GisaidSample, GisaidAccession
 
-from cg.exc import HousekeeperVersionMissingError, FastaSequenceMissingError
+from cg.exc import (
+    HousekeeperFileMissingError,
+)
 
 LOG = logging.getLogger(__name__)
 
@@ -27,96 +31,70 @@ class GisaidAPI:
         self.lims_api: LimsAPI = config.lims_api
         self.status_db: Store = config.status_db
         self.gisaid_submitter: str = config.gisaid.submitter
+        self.upload_password: str = config.gisaid.upload_password
+        self.upload_cid: str = config.gisaid.upload_cid
         self.gisaid_binary: str = config.gisaid.binary_path
+        self.gisaid_log_dir: str = config.gisaid.log_dir
+        self.log_watch: str = config.gisaid.logwatch_email
+        self.email_base_settings = config.email_base_settings
+        self.mutant_root_dir = Path(config.mutant.root)
+
         self.process = Process(binary=self.gisaid_binary)
 
-    def get_new_header(self, samples: List[GisaidSample], old_header: str) -> str:
-        for sample in samples:
-            sample_id = sample.covv_subm_sample_id
-            if old_header.find(sample_id) != -1:
-                return f">{sample.covv_virus_name}\n"
-        raise FastaSequenceMissingError
+    def get_completion_file_from_hk(self, case_id: str) -> File:
+        """Find completon file in Housekeeper and return it"""
 
-    def get_gisaid_fasta(
-        self, gsaid_samples: List[GisaidSample], family_id: str
-    ) -> List[FastaFile]:
-        """Fetch a fasta files form house keeper for batch upload to gisaid"""
-
-        hk_version: hk_models.Version = self.housekeeper_api.last_version(bundle=family_id)
-        if not hk_version:
-            LOG.info("Family ID: %s not found in housekeeper", family_id)
-            raise HousekeeperVersionMissingError
-        fasta_file: str = self.housekeeper_api.files(
-            version=hk_version.id, tags=["consensus"]
-        ).first()
-        gisaid_delivery_fasta = []
-        with open("/Users/maya.brandi/opt/395698.consensus.fa") as handle:
-            fasta_lines = handle.readlines()
-            for line in fasta_lines:
-                if line[0] == ">":
-                    gisaid_delivery_fasta.append(
-                        self.get_new_header(old_header=line, samples=gsaid_samples)
-                    )
-                else:
-                    gisaid_delivery_fasta.append(line)
-
-        return gisaid_delivery_fasta
-
-    def build_gisaid_fasta(
-        self, gsaid_samples: List[GisaidSample], file_name: str, family_id: str
-    ) -> Path:
-        """Concatenates a list of consensus fastq objects"""
-
-        file: Path = Path(file_name)
-        fasta_lines: List[str] = self.get_gisaid_fasta(
-            gsaid_samples=gsaid_samples, family_id=family_id
+        completion_file: Optional[File] = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["komplettering"]
         )
-        with open(file, "w") as write_file_obj:
-            write_file_obj.writelines(fasta_lines)
-        return file
+        if not completion_file:
+            msg = f"completion file missing for bundle {case_id}"
+            raise HousekeeperFileMissingError(message=msg)
+        return completion_file
 
-    def get_sample_row(self, gisaid_sample: GisaidSample) -> List[str]:
-        """Build row for a sample in the batch upload csv."""
+    def get_completion_dataframe(self, completion_file: File) -> pd.DataFrame:
+        """Read completion file in to dataframe, drop duplicates, and return the dataframe"""
+        completion_df = pd.read_csv(completion_file.full_path, index_col=None, header=0)
+        completion_df.drop_duplicates(inplace=True)
+        completion_df = completion_df[completion_df["provnummer"].str.contains("21CS\(|\)|\d{6}")]
+        return completion_df
 
-        sample_dict = gisaid_sample.dict()
-        return [sample_dict.get(header, "") for header in HEADERS]
+    def get_gisaid_sample_list(self, case_id: str) -> List[models.Sample]:
+        """Get list of Sample objects eligeble for upload.
+        The criteria is that the sample reached 20x coverage for >95% bases.
+        The sample will be included in completion file."""
 
-    def get_sample_rows(self, gsaid_samples: List[GisaidSample]) -> List[List[str]]:
-        """Build sample row list for gisad csv file"""
+        completion_file = self.get_completion_file_from_hk(case_id=case_id)
+        completion_df = self.get_completion_dataframe(completion_file=completion_file)
+        sample_names = list(completion_df["provnummer"].unique())
+        return [
+            self.status_db.samples_by_ids(name=sample_name).first() for sample_name in sample_names
+        ]
 
-        sample_rows = []
-        for sample in gsaid_samples:
-            sample_row: List[str] = self.get_sample_row(gisaid_sample=sample)
-            sample_rows.append(sample_row)
-        return sample_rows
+    def get_gisaid_fasta_path(self, case_id: str) -> Path:
+        """Get path to gisaid fasta"""
+        return Path(self.mutant_root_dir, case_id, "results", f"{case_id}.fasta")
 
-    def build_gisaid_csv(self, gsaid_samples: List[GisaidSample], file_name: str) -> Path:
-        """Build batch upload csv."""
+    def get_gisaid_csv_path(self, case_id: str) -> Path:
+        """Get path to gisaid csv"""
+        return Path(self.mutant_root_dir, case_id, "results", f"{case_id}.csv")
 
-        file: Path = Path(file_name)
-        with open(file_name, "w", newline="\n") as gisaid_csv:
-            wr = csv.writer(gisaid_csv, delimiter=",")
-            wr.writerow(HEADERS)
-            sample_rows: List[List[str]] = self.get_sample_rows(gsaid_samples=gsaid_samples)
-            wr.writerows(sample_rows)
-
-        return file
-
-    def get_gsaid_samples(self, family_id: str) -> List[GisaidSample]:
+    def get_gisaid_samples(self, case_id: str) -> List[GisaidSample]:
         """Get list of Gisaid sample objects."""
 
-        samples: List[models.Sample] = self.status_db.get_sequenced_samples(family_id=family_id)
+        samples: List[models.Sample] = self.get_gisaid_sample_list(case_id=case_id)
         gisaid_samples = []
         for sample in samples:
             sample_id: str = sample.internal_id
+            LOG.info(f"Creating GisaidSample for {sample_id}")
             gisaid_sample = GisaidSample(
-                family_id=family_id,
+                case_id=case_id,
                 cg_lims_id=sample_id,
                 covv_subm_sample_id=sample.name,
                 submitter=self.gisaid_submitter,
-                fn=f"{family_id}.fasta",
-                covv_collection_date=self.lims_api.get_sample_attribute(
-                    lims_id=sample_id, key="collection_date"
+                fn=f"{case_id}.fasta",
+                covv_collection_date=str(
+                    self.lims_api.get_sample_attribute(lims_id=sample_id, key="collection_date")
                 ),
                 region=self.lims_api.get_sample_attribute(lims_id=sample_id, key="region"),
                 region_code=self.lims_api.get_sample_attribute(
@@ -132,23 +110,200 @@ class GisaidAPI:
             gisaid_samples.append(gisaid_sample)
         return gisaid_samples
 
-    def upload(self, csv_file: Path, fasta_file: Path) -> None:
-        """Load batch data to GISAID using the gisiad cli."""
+    def create_gisaid_fasta(self, gisaid_samples: List[GisaidSample], case_id: str) -> None:
+        """Writing a new fasta with headers adjusted for gisaid upload_results_to_gisaid"""
 
+        gisaid_fasta_file = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-fasta", case_id]
+        )
+        if gisaid_fasta_file:
+            gisaid_fasta_path = gisaid_fasta_file.full_path
+        else:
+            gisaid_fasta_path: Path = self.get_gisaid_fasta_path(case_id=case_id)
+
+        fasta_lines: List[str] = []
+
+        for sample in gisaid_samples:
+            fasta_file: File = self.housekeeper_api.find_file_in_latest_version(
+                case_id=case_id, tags=[sample.cg_lims_id, "consensus-sample"]
+            )
+            if not fasta_file:
+                raise HousekeeperFileMissingError(
+                    message=f"No fasta file found for sample {sample.cg_lims_id}"
+                )
+            with open(str(fasta_file.full_path)) as handle:
+                for line in handle.readlines():
+                    if line[0] == ">":
+                        fasta_lines.append(f">{sample.covv_virus_name}\n")
+                    else:
+                        fasta_lines.append(line)
+
+        with open(gisaid_fasta_path, "w") as write_file_obj:
+            write_file_obj.writelines(fasta_lines)
+
+        if gisaid_fasta_file:
+            return
+
+        self.housekeeper_api.add_and_include_file_to_latest_version(
+            case_id=case_id, file=gisaid_fasta_path, tags=["gisaid-fasta", case_id]
+        )
+
+    def create_gisaid_csv(self, gisaid_samples: List[GisaidSample], case_id: str) -> None:
+        """Create csv file for gisaid upload"""
+        samples_df = pd.DataFrame(
+            data=[gisaid_sample.dict() for gisaid_sample in gisaid_samples],
+            columns=HEADERS,
+        )
+
+        gisaid_csv_file = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-csv", case_id]
+        )
+        if gisaid_csv_file:
+            LOG.info(f"GISAID CSV for case {case_id} exists, will be replaced")
+            gisaid_csv_path = gisaid_csv_file.full_path
+        else:
+            gisaid_csv_path = self.get_gisaid_csv_path(case_id=case_id)
+        samples_df.to_csv(gisaid_csv_path, sep=",", index=False)
+
+        if gisaid_csv_file:
+            return
+
+        self.housekeeper_api.add_and_include_file_to_latest_version(
+            case_id=case_id, file=gisaid_csv_path, tags=["gisaid-csv", case_id]
+        )
+
+    def create_gisaid_log_file(self, case_id: str) -> None:
+        """Path for gisaid bundle log"""
+        gisaid_log_file = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-log", case_id]
+        )
+        if gisaid_log_file:
+            LOG.info("GISAID log exists in latest bundle in Housekeeper")
+            return
+
+        log_file_path = Path(self.gisaid_log_dir, case_id).with_suffix(".log")
+        if not log_file_path.parent.exists():
+            raise ValueError(f"Gisaid log dir: {self.gisaid_log_dir} doesnt exist")
+        if not log_file_path.exists():
+            log_file_path.touch()
+        self.housekeeper_api.add_and_include_file_to_latest_version(
+            case_id=case_id, file=log_file_path, tags=["gisaid-log", case_id]
+        )
+
+    def create_gisaid_files_in_housekeeper(self, case_id: str) -> None:
+        """Create all gisaid files in Housekeeper, if needed."""
+        gisaid_samples = self.get_gisaid_samples(case_id=case_id)
+        self.create_gisaid_csv(gisaid_samples=gisaid_samples, case_id=case_id)
+        self.create_gisaid_fasta(gisaid_samples=gisaid_samples, case_id=case_id)
+        self.create_gisaid_log_file(case_id=case_id)
+
+    def authenticate_gisaid(self):
         load_call: list = [
             "CoV",
-            "upload",
-            "--csv",
-            str(csv_file.absolute()),
-            "--fasta",
-            str(fasta_file.absolute()),
+            "authenticate",
+            "--cid",
+            self.upload_cid,
+            "--user",
+            self.gisaid_submitter,
+            "--pass",
+            self.upload_password,
         ]
         self.process.run_command(parameters=load_call)
 
+    def upload_results_to_gisaid(self, case_id: str) -> None:
+        """Load batch data to GISAID using the gisiad cli."""
+
+        temp_log_file = tempfile.NamedTemporaryFile(
+            dir=self.gisaid_log_dir, mode="w+", delete=False
+        )
+        gisaid_csv_path = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-csv", case_id]
+        ).full_path
+
+        gisaid_fasta_path = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-fasta", case_id]
+        ).full_path
+
+        gisaid_log_path = self.housekeeper_api.find_file_in_latest_version(
+            case_id=case_id, tags=["gisaid-log", case_id]
+        ).full_path
+
+        self.authenticate_gisaid()
+        load_call: list = [
+            "--logfile",
+            temp_log_file.name,
+            "CoV",
+            "upload",
+            "--csv",
+            gisaid_csv_path,
+            "--fasta",
+            gisaid_fasta_path,
+        ]
+        self.process.run_command(parameters=load_call)
+        self.append_log(temp_log=Path(temp_log_file.name), gisaid_log=Path(gisaid_log_path))
+        temp_log_file.close()
         if self.process.stderr:
             LOG.info(f"gisaid stderr:\n{self.process.stderr}")
         if self.process.stdout:
             LOG.info(f"gisaid stdout:\n{self.process.stdout}")
 
-    def __str__(self):
-        return f"GisaidAPI(dry_run: {self.dry_run})"
+    def append_log(self, temp_log: Path, gisaid_log: Path) -> None:
+        """Appends temp log to gisaid log and delete temp file"""
+
+        with open(str(temp_log.absolute()), "r") as open_temp_log:
+            new_log_data: List = json.load(open_temp_log)
+            if gisaid_log.stat().st_size != 0:
+                with open(str(gisaid_log.absolute()), "r") as open_gisaid_log:
+                    old_log_data: List = json.load(open_gisaid_log)
+                    new_log_data.extend(old_log_data)
+
+        with open(str(gisaid_log.absolute()), "w") as open_gisaid_log:
+            json.dump(new_log_data, open_gisaid_log)
+        temp_log.unlink()
+
+    def get_accession_numbers(self, case_id: str) -> Dict[str, str]:
+        """Parse accession numbers and sample ids from log file"""
+
+        LOG.info("Parsing accesion numbers from log file")
+        accession_numbers = {}
+        log_file = Path(
+            self.housekeeper_api.find_file_in_latest_version(
+                case_id=case_id, tags=["gisaid-log", case_id]
+            ).full_path
+        )
+        if log_file.stat().st_size != 0:
+            with open(str(log_file.absolute())) as log_file:
+                log_data = json.load(log_file)
+                for log in log_data:
+                    if log.get("code") != "epi_isl_id":
+                        continue
+                    accession_obj = GisaidAccession(log_message=log.get("msg"))
+                    accession_numbers[accession_obj.sample_id] = accession_obj.accession_nr
+        return accession_numbers
+
+    def update_completion_file(self, case_id: str) -> None:
+        """Update completion file with accession numbers"""
+        completion_file = self.get_completion_file_from_hk(case_id=case_id)
+        accession_dict = self.get_accession_numbers(case_id=case_id)
+        completion_df = self.get_completion_dataframe(completion_file=completion_file)
+        completion_df["GISAID_accession"] = completion_df["provnummer"].apply(
+            lambda x: accession_dict[x]
+        )
+        completion_df.to_csv(
+            completion_file.full_path,
+            sep=",",
+            index=False,
+        )
+
+    def upload(self, case_id: str) -> None:
+        """Uploading results to gisaid and saving the accession numbers in completion file"""
+
+        completion_file = self.get_completion_file_from_hk(case_id=case_id)
+        completion_df = self.get_completion_dataframe(completion_file=completion_file)
+        if len(completion_df["GISAID_accession"].dropna()) == len(completion_df["provnummer"]):
+            LOG.info("All samples already uploaded")
+            return
+
+        self.create_gisaid_files_in_housekeeper(case_id=case_id)
+        self.upload_results_to_gisaid(case_id=case_id)
+        self.update_completion_file(case_id=case_id)